@@ -1,22 +1,26 @@
 module sigs.k8s.io/cluster-api-provider-kubevirt
 
-go 1.17
+go 1.18
 
 require (
-	github.com/go-logr/logr v1.2.0
+	github.com/go-logr/logr v1.2.3
+	github.com/golang/glog v1.0.0
 	github.com/golang/mock v1.6.0
-	github.com/onsi/ginkgo/v2 v2.1.3
-	github.com/onsi/gomega v1.17.0
+	github.com/onsi/ginkgo/v2 v2.1.4
+	github.com/onsi/gomega v1.19.0
 	github.com/pkg/errors v0.9.1
 	github.com/spf13/cobra v1.3.0
 	github.com/spf13/pflag v1.0.5
 	golang.org/x/crypto v0.0.0-20211117183948-ae814b36b871
-	k8s.io/api v0.23.1
-	k8s.io/apimachinery v0.23.1
-	k8s.io/client-go v0.23.1
+	k8s.io/api v0.23.5
+	k8s.io/apimachinery v0.23.5
+	k8s.io/client-go v12.0.0+incompatible
 	k8s.io/component-base v0.23.1
-	k8s.io/klog/v2 v2.30.0
-	kubevirt.io/api v0.51.0
+	k8s.io/klog/v2 v2.40.1
+	k8s.io/kubectl v0.22.2
+	k8s.io/utils v0.0.0-20211116205334-6203023598ed
+	kubevirt.io/api v0.53.0
+	kubevirt.io/client-go v0.0.0-00010101000000-000000000000
 	sigs.k8s.io/cluster-api v0.3.11-0.20210525210043-6c7878e7b4a9
 	sigs.k8s.io/controller-runtime v0.11.0-beta.0.0.20211110210527-619e6b92dab9
 	sigs.k8s.io/kind v0.11.0
